Revert "Merge branch 'main' into DEV_JSB"
This reverts commit e7f2169e20fe7358e2eb1ca1ab1bb7be20bbaf1a, reversing
changes made to f45d1c7b490723cb7b6479fe8a9f0ef313a45762.

+++ DimensionArena/Assets/02.Scripts/06.Item/ItemBox.cs
@@ -127,7 +127,7 @@ public class ItemBox : MonoBehaviourPun
     public void HpDecrease_KnockBack(float damage, string owner)
     {
         if (!PhotonNetwork.OfflineMode)
-            photonView.RPC(nameof(HpDecrease), RpcTarget.All, (int)damage, owner);
+            photonView.RPC(nameof(HpDecrease), RpcTarget.All, damage, owner);
         else
             HpDecrease((int)damage, owner);
     }
